release: foldable sidbar
