+++ docs/index.html
@@ -5,8 +5,8 @@
     <link rel="icon" href="/assets/sakurajima.44cdfb32.jpeg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Tachish's Blog</title>
-    <script type="module" crossorigin src="/assets/index.f8e2f231.js"></script>
-    <link rel="stylesheet" href="/assets/index.4e484a90.css">
+    <script type="module" crossorigin src="/assets/index.4d13d089.js"></script>
+    <link rel="stylesheet" href="/assets/index.d94df9e2.css">
   </head>
   <body>
     <div id="app"></div>
